No idea of the way the UserBar sometime work properly sometime not, I'm wuitting that idea after this commit.

+++ server/src/http/routes/delete-comment.ts
@@ -1,14 +1,27 @@
 import { FastifyInstance } from "fastify";
 import { z } from "zod";
 import { prisma } from "../../lib/prisma";
+import { verifyToken } from "../token";
+import { postsPubSub } from "../../utils/posts-pub-sub";
 
 export async function deleteComment(app: FastifyInstance) {
-    app.post('/delete/comment/:commentId', async (request, reply) => {
-        const deleteComment = z.object({
+    app.delete('/comment/:commentId', async (request, reply) => {
+        const zCommentId = z.object({
             commentId: z.string(),
-            userId: z.string(),
         })
-        const { commentId, userId } = deleteComment.parse(request.params)
+        const zCommentBody = z.object({
+            userId: z.string(),
+            token: z.string(),
+        })
+        const { commentId } = zCommentId.parse(request.params)
+        const { userId, token } = zCommentBody.parse(request.query)
+        const verifiedToken = verifyToken(token);
+        if (!verifiedToken.valid) {
+            return reply.status(400).send({ message: "Not authorized" });
+        }
+        if (!commentId) {
+            return reply.status(404).send({ message: "Error deleting the comment" });
+        }
         await prisma.comment.deleteMany({
             where: {
                 parentCommentId: commentId,
@@ -21,6 +34,7 @@ export async function deleteComment(app: FastifyInstance) {
                 userId: userId,
             },
         })
+        postsPubSub.publish('asf', { action: 'delete', type: 'comment', data: { id: commentId, userId } })
         return reply.status(201).send()
     })
 }
